[Refactor] FillerWordDetail 코드 가독성 개선 #4
