+++ highpitch/Presentation/MainWindow/ProjectView/PracticeView/FeedbackChartView/FeedbackChartView.swift
@@ -8,6 +8,9 @@
 import SwiftUI
 
 struct FeedbackChartView: View {
+    @Environment(PracticeViewStore.self)
+    var viewStore
+    
     @State
     var practice: PracticeModel
     @State
@@ -19,62 +22,10 @@ struct FeedbackChartView: View {
         VStack(spacing: 0) {
             header
             ScrollView {
-                VStack(spacing: 0) {
-                    HStack(alignment: .top, spacing: .HPSpacing.xxxxsmall) {
-                        Text("이번 연습에서 사용한 습관어")
-                            .systemFont(.title)
-                            .foregroundStyle(Color.HPTextStyle.darker)
-                        tooltipFillerWord
-                    }
-                    .frame(
-                        maxWidth: .infinity,
-                        alignment: .leading
-                    )
-                    .padding(.bottom, .HPSpacing.xsmall)
-                    UsagePercentChart(practiceModel: $practice)
-                    UsageTopTierChart(summary: practice.summary)
-                    if (practice.summary.fillerWordCount > 0) {
-                        FillerWordDetail(practiceModel: $practice)
-                            .padding(.bottom, .HPSpacing.medium)
-                    }
-                }
-                .frame(
-                    maxWidth: 720,
-                    maxHeight: .infinity,
-                    alignment: .leading
-                )
+                fillerWordChartContainer
                 Divider()
                     .padding(.trailing, .HPSpacing.medium)
-                VStack(spacing: 0) {
-                    HStack(alignment: .top, spacing: .HPSpacing.xxxxsmall) {
-                        Text("이번 연습에서의 말 빠르기")
-                            .systemFont(.title)
-                            .foregroundStyle(Color.HPTextStyle.darker)
-                        tooltipSpeed
-                    }
-                    .frame(
-                        maxWidth: .infinity,
-                        maxHeight: .infinity,
-                        alignment: .leading
-                    )
-                    SpeedAverageChart(
-                        sentences: practice.sentences,
-                        practice: practice,
-                        epmRange: getEpmRange()
-                    )
-                }
-                .padding(.top, .HPSpacing.medium)
-                .frame(
-                    maxWidth: 720,
-                    maxHeight: .infinity,
-                    alignment: .leading
-                )
-                if (!practice.summary.fastSentenceIndex.isEmpty) {
-                    FastSentReplay(practice: practice)
-                }
-                if (!practice.summary.slowSentenceIndex.isEmpty) {
-                    SlowSentReplay(practice: practice)
-                }
+                speedChartContainer
             }
             .padding(.bottom, 100)
             .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .trailing)
@@ -85,15 +36,6 @@ struct FeedbackChartView: View {
 }
 
 extension FeedbackChartView {
-    
-    func getEpmRange() -> [Double] {
-        let sortedSentences = practice.sentences.sorted(by: { $0.epmValue < $1.epmValue })
-        return [
-            sortedSentences.first!.epmValue,
-            sortedSentences.last!.epmValue
-        ]
-    }
-    
     @ViewBuilder
     private var header: some View {
         VStack(spacing: 0) {
@@ -103,6 +45,70 @@ extension FeedbackChartView {
         }
         .padding(.bottom, .HPSpacing.xsmall)
         .frame(maxWidth: 740, alignment: .leading)        
+    }
+    
+    @ViewBuilder
+    private var fillerWordChartContainer: some View {
+        VStack(spacing: 0) {
+            HStack(alignment: .top, spacing: .HPSpacing.xxxxsmall) {
+                Text("이번 연습에서 사용한 습관어")
+                    .systemFont(.title)
+                    .foregroundStyle(Color.HPTextStyle.darker)
+                tooltipFillerWord
+            }
+            .frame(
+                maxWidth: .infinity,
+                alignment: .leading
+            )
+            .padding(.bottom, .HPSpacing.xsmall)
+            UsagePercentChart(practiceModel: $practice)
+            UsageTopTierChart(summary: practice.summary)
+            if viewStore.hasFillerWord() {
+                FillerWordDetail(fillerWords: viewStore.getSortedFillerWord())
+                    .padding(.bottom, .HPSpacing.medium)
+            }
+        }
+        .frame(
+            maxWidth: 720,
+            maxHeight: .infinity,
+            alignment: .leading
+        )
+    }
+    
+    @ViewBuilder
+    private var speedChartContainer: some View {
+        VStack(spacing: 0) {
+            VStack(spacing: 0) {
+                HStack(alignment: .top, spacing: .HPSpacing.xxxxsmall) {
+                    Text("이번 연습에서의 말 빠르기")
+                        .systemFont(.title)
+                        .foregroundStyle(Color.HPTextStyle.darker)
+                    tooltipSpeed
+                }
+                .frame(
+                    maxWidth: .infinity,
+                    maxHeight: .infinity,
+                    alignment: .leading
+                )
+                SpeedAverageChart(
+                    sentences: viewStore.practice.sentences,
+                    practice: viewStore.practice,
+                    epmRange: viewStore.getEpmRange()
+                )
+            }
+            .padding(.top, .HPSpacing.medium)
+            .frame(
+                maxWidth: 720,
+                maxHeight: .infinity,
+                alignment: .leading
+            )
+            if (!practice.summary.fastSentenceIndex.isEmpty) {
+                FastSentReplay(practice: practice)
+            }
+            if (!practice.summary.slowSentenceIndex.isEmpty) {
+                SlowSentReplay(practice: practice)
+            }
+        }
     }
     
     @ViewBuilder
